Factory and Seeder to insert faker viviendas in migrations

+++ inmobiliaria/database/migrations/2020_05_19_094119_create_viviendas_table.php
@@ -15,7 +15,7 @@ class CreateViviendasTable extends Migration
     {
         Schema::create('viviendas', function (Blueprint $table) {
             $table->id();
-            $table->integer('precio');
+            $table->decimal('precio', 12);
             $table->integer('superficie');
             $table->integer('habitaciones');
             $table->date('construcción');
